refactor calculate-iob to be more testable/reusable
